add the posibility of specifying a url prefix
this commit implements a prefix for urls, which allows easily deploying
dpaste into a "subdirectory". It adds a url_prefix to be read by default
from the environment which makes it especially suitable for using for
docker containers. For example, running

docker run -e PREFIX_URL=dpaste/ -p 8000:8000 barttc/dpaste

will make dpaste available on

http://localhost:8000/dpaste/

instead of the default http://localhost:8000

This is specially useful if you want to proxy dpaste from another host
and let it live in a sub-url. Without this commit all urls would be
incorrectly generated in this case.